Performed a whole lot of "house keeping" tasks
Highlights include:
* Added prelimiary ANT support and eventual CI support
* All project static resources are now centralized to a single static
resource object called "apexuml".  All of the contents are managed in
the "/resource-bundles/apexuml" folder.  Assembly of the static resource
bundle is automated by executing the ANT target
"package-and-install-all-static-resources".  This will update the
apexuml.resource file in the project folder
* The old Umlcanvas.js file has been moved to the ApexNavigator VF page.
This will allow us to take advantage of the UMLController Apex class and
other VF functions (e.g. urlfor()).  The ApexNavigator page is inserted
into the uml VF page, but kept in separate file for ease of editing.
* uml.page is now responsible for loading all umlcanvas related
javascript libraries
* project package.xml file now explicitly specifies all project
metadata.
* Basic Javascript wiring from ApexNavigator javascript object is
complete.  UMLService now needs to be fleshed out.