Work relates to issue #6
Temporarily am trapping the CalloutException in the ToolingAPI and
throwing a ToolingAPIException.

The UmlService is now trapping the ToolingAPI.ToolingAPIException in
certain places and creates a new UmlServiceException to throw to the
client tier (i.e. UmlController).

This is a temporary solution until the debate about expanding the
ToolingAPI.ToolingAPIException (see comments on this issue) is
concluded.  

These changes should remain on this branch until debate is comcluded.