Resolves #3
@afawcett - I implemented a workaround here.

ToolingAPI.SymbolTable mods
* Added boolean isInterface

UmlService mods
* Added new private method augumentSymbolTable(ToolingApi.SymbolTable).
This method will query the ApexClass directly to determine if the
ApexClass is an interface or not.
*** This method is not bulkified yet.
*** This method assumes the declaration " interface " will be on the
same line as the name of the ApexClass
*** This method has not been tested on "inner class interfaces" yet.

uml.component mods
* constructUMLClass : function (classSymbolTable, parentClassName) has
been modified to use new SymbolTable "isInterface" parameter.  If the
SymbolTable.isInterface is true, then the umlcanvas will construct an
new UmlCanvas.Interface object verses a new UmlCanvas.Class object.

The augumentSymbolTable method was added to the UmlService class and not
to the ToolingAPI deliberately.  At this time, we do not know when or
how Salesforce will update the Tooling API to designate if an ApexClass
is an interface.  Until that information is learned, we should limit
this change to this apex-umlcanvas project only.  This way, if we find
out that we need to implement this feature in a different manner, it
only affects this project and not other projects that may depend on
ToolingAPI service.

cc: @dancinlama and @alan-morey