Resolves #12
This should resolve the issue.

Instead of relying on the apex:repeat method to produce the class list,
take the class list and serialize it.  That string is placed on the
webpage and let jQuery parse the JSON and populate the list elements for
the DOM object "classlist".

This approach should allow us to get past the 1000 element list limit.